requirements.txt
Babel==2.12.1
py-moneyed==3.0
typing_extensions==4.6.2
